move repos related actions back to app container
