+++ app/containers/App/reducer.js
@@ -10,16 +10,42 @@
  *   return state.set('yourStateVariable', true);
  */
 
+import {
+  LOAD_REPOS_SUCCESS,
+  LOAD_REPOS,
+  LOAD_REPOS_ERROR,
+} from './constants';
 import { fromJS } from 'immutable';
 
 // The initial state of the App
-const initialState = fromJS({});
+const initialState = fromJS({
+  loading: false,
+  error: false,
+  currentUser: false,
+  userData: fromJS({
+    repositories: false,
+  }),
+});
 
-function globalReducer(state = initialState, action) {
+function homeReducer(state = initialState, action) {
   switch (action.type) {
+    case LOAD_REPOS:
+      return state
+        .set('loading', 'true')
+        .set('error', false)
+        .setIn(['userData', 'repositories'], false);
+    case LOAD_REPOS_SUCCESS:
+      return state
+        .setIn(['userData', 'repositories'], action.repos)
+        .set('loading', false)
+        .set('currentUser', action.username);
+    case LOAD_REPOS_ERROR:
+      return state
+        .set('error', action.error)
+        .set('loading', false);
     default:
       return state;
   }
 }
 
-export default globalReducer;
+export default homeReducer;
