+++ app/containers/HomePage/actions.js
@@ -17,9 +17,6 @@
 
 import {
   CHANGE_USERNAME,
-  LOAD_REPOS,
-  LOAD_REPOS_SUCCESS,
-  LOAD_REPOS_ERROR,
 } from './constants';
 
 /**
@@ -33,44 +30,5 @@ export function changeUsername(name) {
   return {
     type: CHANGE_USERNAME,
     name,
-  };
-}
-
-/**
- * Load the repositories, this action starts the getGithubData saga
- *
- * @return {object} An action object with a type of LOAD_REPOS
- */
-export function loadRepos() {
-  return {
-    type: LOAD_REPOS,
-  };
-}
-
-/**
- * Dispatched when the repositories are loaded by the getGithubData saga
- *
- * @param  {array} repos The repository data
- *
- * @return {object}      An action object with a type of LOAD_REPOS_SUCCESS passing the repos
- */
-export function reposLoaded(repos) {
-  return {
-    type: LOAD_REPOS_SUCCESS,
-    repos,
-  };
-}
-
-/**
- * Dispatched when loading the repositoreis fails
- *
- * @param  {object} error The error
- *
- * @return {object}       An action object with a type of LOAD_REPOS_ERROR passing the error
- */
-export function repoLoadingError(error) {
-  return {
-    type: LOAD_REPOS_ERROR,
-    error,
   };
 }
