+++ app/containers/HomePage/reducer.js
@@ -12,21 +12,12 @@
 
 import {
   CHANGE_USERNAME,
-  LOAD_REPOS_SUCCESS,
-  LOAD_REPOS,
-  LOAD_REPOS_ERROR,
 } from './constants';
 import { fromJS } from 'immutable';
 
 // The initial state of the App
 const initialState = fromJS({
-  loading: false,
-  error: false,
-  currentUser: false,
-  userData: fromJS({
-    repositories: false,
-    username: '',
-  }),
+  username: '',
 });
 
 function homeReducer(state = initialState, action) {
@@ -34,21 +25,7 @@ function homeReducer(state = initialState, action) {
     case CHANGE_USERNAME:
       // Delete prefixed '@' from the github username
       return state
-        .setIn(['userData', 'username'], action.name.replace(/@/gi, ''));
-    case LOAD_REPOS:
-      return state
-        .set('loading', 'true')
-        .set('error', false)
-        .setIn(['userData', 'repositories'], false);
-    case LOAD_REPOS_SUCCESS:
-      return state
-        .setIn(['userData', 'repositories'], action.repos)
-        .set('loading', false)
-        .set('currentUser', state.getIn(['userData', 'username']));
-    case LOAD_REPOS_ERROR:
-      return state
-        .set('error', action.error)
-        .set('loading', false);
+        .set('username', action.name.replace(/@/gi, ''));
     default:
       return state;
   }
