+++ app/containers/HomePage/tests/actions.test.js
@@ -2,16 +2,10 @@ import expect from 'expect';
 
 import {
   CHANGE_USERNAME,
-  LOAD_REPOS,
-  LOAD_REPOS_SUCCESS,
-  LOAD_REPOS_ERROR,
 } from '../constants';
 
 import {
   changeUsername,
-  loadRepos,
-  reposLoaded,
-  repoLoadingError,
 } from '../actions';
 
 describe('Home Actions', () => {
@@ -24,42 +18,6 @@ describe('Home Actions', () => {
       };
 
       expect(changeUsername(fixture)).toEqual(expectedResult);
-    });
-  });
-
-  describe('loadRepos', () => {
-    it('should return the correct type', () => {
-      const expectedResult = {
-        type: LOAD_REPOS,
-      };
-
-      expect(loadRepos()).toEqual(expectedResult);
-    });
-  });
-
-  describe('reposLoaded', () => {
-    it('should return the correct type and the passed repos', () => {
-      const fixture = ['Test'];
-      const expectedResult = {
-        type: LOAD_REPOS_SUCCESS,
-        repos: fixture,
-      };
-
-      expect(reposLoaded(fixture)).toEqual(expectedResult);
-    });
-  });
-
-  describe('repoLoadingError', () => {
-    it('should return the correct type and the error', () => {
-      const fixture = {
-        msg: 'Something went wrong!',
-      };
-      const expectedResult = {
-        type: LOAD_REPOS_ERROR,
-        error: fixture,
-      };
-
-      expect(repoLoadingError(fixture)).toEqual(expectedResult);
     });
   });
 });
