+++ app/containers/HomePage/tests/reducer.test.js
@@ -2,9 +2,6 @@ import expect from 'expect';
 import homeReducer from '../reducer';
 import {
   changeUsername,
-  loadRepos,
-  reposLoaded,
-  repoLoadingError,
 } from '../actions';
 import { fromJS } from 'immutable';
 
@@ -12,13 +9,7 @@ describe('homeReducer', () => {
   let state;
   beforeEach(() => {
     state = fromJS({
-      loading: false,
-      error: false,
-      currentUser: false,
-      userData: fromJS({
-        repositories: false,
-        username: '',
-      }),
+      username: '',
     });
   });
 
@@ -29,40 +20,8 @@ describe('homeReducer', () => {
 
   it('should handle the changeUsername action correctly', () => {
     const fixture = 'mxstbr';
-    const expectedResult = state.setIn(['userData', 'username'], fixture);
+    const expectedResult = state.set('username', fixture);
 
     expect(homeReducer(state, changeUsername(fixture))).toEqual(expectedResult);
-  });
-
-  it('should handle the loadRepos action correctly', () => {
-    const expectedResult = state
-      .set('loading', 'true')
-      .set('error', false)
-      .setIn(['userData', 'repositories'], false);
-
-    expect(homeReducer(state, loadRepos())).toEqual(expectedResult);
-  });
-
-  it('should handle the reposLoaded action correctly', () => {
-    const fixture = [{
-      name: 'My Repo',
-    }];
-    const expectedResult = state
-      .setIn(['userData', 'repositories'], fixture)
-      .set('loading', false)
-      .set('currentUser', state.getIn(['userData', 'username']));
-
-    expect(homeReducer(state, reposLoaded(fixture))).toEqual(expectedResult);
-  });
-
-  it('should handle the repoLoadingError action correctly', () => {
-    const fixture = {
-      msg: 'Not found',
-    };
-    const expectedResult = state
-      .set('error', fixture)
-      .set('loading', false);
-
-    expect(homeReducer(state, repoLoadingError(fixture))).toEqual(expectedResult);
   });
 });
