+++ app/sagas/tests/getGithubData.test.js
@@ -8,15 +8,16 @@ import { take, call, put, select } from 'redux-saga/effects';
 import { getGithubData } from '../getGithubData.saga';
 import {
   LOAD_REPOS,
-} from 'HomePage/constants';
+} from 'App/constants';
 import {
   reposLoaded,
   repoLoadingError,
-} from 'HomePage/actions';
+} from 'App/actions';
 import request from '../../utils/request';
 import usernameSelector from 'usernameSelector';
 
 const generator = getGithubData();
+const username = 'mxstbr';
 
 describe('getGithubData Saga', () => {
   // We have to test twice, once for a successful load and once for an unsuccessful one
@@ -24,7 +25,6 @@ describe('getGithubData Saga', () => {
   beforeEach(() => {
     expect(generator.next().value).toEqual(take(LOAD_REPOS));
     expect(generator.next().value).toEqual(select(usernameSelector));
-    const username = 'mxstbr';
     const requestURL = 'https://api.github.com/users/' + username + '/repos?type=all&sort=updated';
     expect(generator.next(username).value).toEqual(call(request, requestURL));
   });
@@ -37,7 +37,7 @@ describe('getGithubData Saga', () => {
         name: 'Second repo',
       }],
     };
-    expect(generator.next(response).value).toEqual(put(reposLoaded(response.data)));
+    expect(generator.next(response).value).toEqual(put(reposLoaded(response.data, username)));
   });
 
   it('should call the repoLoadingError action if the response errors', () => {
