+++ app/selectors/tests/errorSelector.test.js
@@ -10,7 +10,7 @@ describe('errorSelector', () => {
   it('should select the error', () => {
     const error = 404;
     const mockedState = fromJS({
-      home: {
+      global: {
         error,
       },
     });
